% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAIS.R
\name{addAIS}
\alias{addAIS}
\title{Add AIS Data to Dataframe}
\usage{
addAIS(
  x,
  ais,
  interpType = c("all", "close", "none"),
  interpTime = 0,
  interpCols = NULL
)
}
\arguments{
\item{x}{a dataframe with \code{UTC}, \code{Latitude}, and \code{Longitude}
columns}

\item{ais}{AIS data created using the \link{readLocalAIS} function}

\item{interpType}{one of \code{c('all', 'close', 'none')}, the type
of time interpolation to apply to \code{x}. Often the time scale of points
in \code{x} is much longer than the points in \code{ais}, which can result
in awkward looking AIS paths. \code{'all'} will interpolate all points in
\code{x} to a smaller timescale. \code{'close'} will interpolate only
time ranges in \code{ais} marked as \code{inDist} by \link{readLocalAIS}.
\code{'none'} will apply no interpolation}

\item{interpTime}{time (seconds) between new \code{UTC} points. If
\code{0} (default), no interpolation will be done}

\item{interpCols}{names of any extra columns to interpolate (other than
\code{Latitude} and \code{Longitude})}
}
\value{
a dataframe with AIS data added, will contain more rows than \code{x}
if \code{ais} has more than one vessel. If any interpolation is applied,
any non-constant columns not specified to \code{interpCols} will be removed
}
\description{
Adds matching AIS data downloaded from Marine Cadastre
to a dataframe containing location information
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
