% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSoundscapeData.R
\name{binSoundscapeData}
\alias{binSoundscapeData}
\title{Summarise Soundscape Data by Time Bin}
\usage{
binSoundscapeData(x, bin = "1hour", FUN = median)
}
\arguments{
\item{x}{a data.frame of soundscape metric data read in with
\link{loadSoundscapeData}}

\item{bin}{amount of time to bin data by, format can
be "#Unit" e.g. \code{'2hour'} or \code{'1day'}}

\item{FUN}{summary function to apply to data in each time bin}
}
\value{
a summarised version of the input data.frame \code{x}
}
\description{
Bins soundscape measurements by a unit of time
and summarises them using a function (usually the median)
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
