% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDailyLTSAReview.R
\name{runDailyLTSAReview}
\alias{runDailyLTSAReview}
\title{Run Daily Soundscape LTSA Data Review Shiny App}
\usage{
runDailyLTSAReview(file, plotQuality = FALSE)
}
\arguments{
\item{file}{either a folder containing NetCDF soundscape files or a
single file}

\item{plotQuality}{logical flag whether or not to plot the matrix of
data quality flags. This greatly increases the load time for each file.}
}
\value{
nothing, just runs the app
}
\description{
Run Daily Soundscape LTSA Data Review Shiny App
}
\details{
Launches a Shiny app that allows users to manually review LTSA
style plots of soundscape NetCDF data. Users can draw annotations
on the LTSA plot and export these annotations as a CSV file. Users can
additionally mark selected areas with the 1-4 standard quality flags.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
