% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDetLoadHelper.R
\name{runDetLoadHelper}
\alias{runDetLoadHelper}
\title{Run Detection Data Loading Helper App}
\usage{
runDetLoadHelper(data = NULL)
}
\arguments{
\item{data}{file path to a CSV file containing detection data}
}
\description{
Runs a Shiny app that allows users to interactively
find the proper parameters to load a detection dataset using
\link{loadDetectionData}
}
\examples{
if(interactive()) {
  detFile <- system.file('extdata/detectionExample.csv', package='PAMscapes')
  runDetLoadHelper(detFile)
}

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
