% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQAQCReview.R
\name{runQAQCReview}
\alias{runQAQCReview}
\title{Run QAQC Review Shiny App}
\usage{
runQAQCReview(data, issue = NULL, freqLims = c(30, Inf))
}
\arguments{
\item{data}{output from \link{evaluateDeployment} either as a dataframe
or path to a CSV file containing the same data}

\item{issue}{if not \code{NULL}, a dataframe or path to CSV file of
issue logs previously created with this function}

\item{freqLims}{frequency limits (Hz) for plotting}
}
\value{
invisibly \code{data}
}
\description{
Launches a shiny app that allows users to review QAQC outputs
created by \link{evaluateDeployment} and interactively mark potential
problems to investigate. Potential problems can be saved in an "Issues Log"
that can be downloaded as a CSV file.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
