% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDataQuality.R
\name{addDataQuality}
\alias{addDataQuality}
\alias{formatDQDf}
\title{Add Data Quality Matrix to PyPam NetCDF}
\usage{
addDataQuality(
  x,
  annotation = NULL,
  platformColumn = NULL,
  updateUnknown = FALSE,
  preview = FALSE,
  plot = FALSE,
  reset = FALSE,
  verbose = FALSE
)

formatDQDf(annotation)
}
\arguments{
\item{x}{one or more NetCDF files or a folder containing NetCDF files.
If \code{x} is a folder, all files with extension ".nc" will be
listed from the first level of the folder (non-recursive)}

\item{annotation}{a dataframe of annotations specifying the time-frequency
bounds and quality flags to be changed. Columns for time bounds must
be called \code{start} and \code{end}, must be in UTC timezone, and
either POSIXct format or in YYYY-MM-DD HH:MM:SS. Columns for frequency
bounds must be called \code{freqMin} and \code{freqMax} and have numeric
frequency values in Hz. Column for quality code must be called
\code{quality} and have numeric values from 1,2,3,4. Can also have
additional column \code{file} specifying that a given annotation only
applies to a specific file, or a column to match the "platform" attribute
of the NetCDF file (see \code{platformColumn})}

\item{platformColumn}{NetCDF files should contain a "platform" global attribute,
this can be used to specify that certain rows of \code{annotation} only apply
to certain files in \code{x}. This value specifies the name of the column in
\code{annotation} should be used (e.g. "deployment_code" means
\code{annotation$deployment_code} contains values that will match the
"platform" global attribute)}

\item{updateUnknown}{if \code{FALSE}, then existing quality flags of "2"
will not be changed, otherwise flags of "2" will be updated to this value
(must be one of 1, 3, or 4)}

\item{preview}{if \code{TRUE}, then no changes are applied to the NetCDF
files, instead only plots are shown. It is recommended to run with
\code{preview=TRUE} first to check that annotations appear to be applied
properly}

\item{plot}{if \code{TRUE}, a plot of the quality matrix is shown. If
\code{"new"}, the plot is only shown if any annotations have been
applied to the file. If \code{FALSE} no plot is created}

\item{reset}{if \code{TRUE}, all quality flag flags are reset to a
value of "2" before applying further annotations. Can be useful if
incorrect annotations were previously applied. If \code{reset=TRUE}
and \code{annotation=NULL}, then all quality code values will be "2"}

\item{verbose}{logical flag to show or hide some messages}
}
\value{
\code{TRUE} or \code{FALSE} to indicate whether or not any
annotations were applied. If \code{x} is multiple files or a folder of
files, the returned vector has names equal to the file names
}
\description{
Add or modify the existing data quality matrix data
in a PyPam (or similar) hybrid millidecade NetCDF output. This
data adds or modifies the "quality_flag" variable of a NetCDF
file, which is a matrix of numbers 1,2,3, or 4 that is the same
dimensions as the hybrid millidecade outputs. The input NetCDF files
are modified unless \code{preview=TRUE}
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
