% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMantaNc.R
\name{loadMantaNc}
\alias{loadMantaNc}
\title{Load MANTA NetCDF File}
\usage{
loadMantaNc(x)
}
\arguments{
\item{x}{path to .nc file}
}
\value{
a dataframe with first column UTC and other columns
named HMD_Frequency
}
\description{
Reads in hybrid millidecade data from a MANTA
NetCDF output file and formats it into the dataframe format
required for use in other PAMscapes functions
}
\examples{
# no sample NetCDF provided (too large)
\donttest{
manta <- loadMantaNc('MANTA.nc')
}

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
