% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGFS.R
\name{matchGFS}
\alias{matchGFS}
\title{Match GFS Environmental Data}
\usage{
matchGFS(x)
}
\arguments{
\item{x}{a dataframe with columns \code{UTC}, \code{Latitude} and
\code{Longitude} to add environmental data to}
}
\value{
a dataframe with wind and precipitation rate columns added
}
\description{
Downloads and matches wind and precipitation data
from the Global Forecast System (GFS) weather model. Data is
downloaded from the UCAR data server
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
