% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPSD.R
\name{plotPSD}
\alias{plotPSD}
\title{Plot Power Spectral Density}
\usage{
plotPSD(
  x,
  style = c("quantile", "density"),
  scale = c("log", "linear"),
  q = 0.5,
  color = "black",
  freqRange = NULL,
  dbRange = NULL,
  dbInt = 1,
  units = "dB re: 1uPa",
  cmap = viridis_pal()(25),
  title = NULL
)
}
\arguments{
\item{x}{a dataframe}

\item{style}{character specifying plot style to create, either
"quantile", "density", or a vector with both}

\item{scale}{scale to use for frequency axis, one of "log" or "linear"}

\item{q}{quantile to plot}

\item{color}{color for quantile}

\item{freqRange}{range of frequencies to plot}

\item{dbRange}{range of dB values to plot}

\item{dbInt}{bin interval size for density plot}

\item{units}{units for dB axis of plot}

\item{cmap}{color map to use for density plot}

\item{title}{optional title for plot}
}
\value{
a ggplot object
}
\description{
Plots the distribution of summarised sound levels
across frequency, either as lines of quantile levels or a
heatmap showing the full distribution
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
