% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQAQC.R
\name{plotQAQCLevel}
\alias{plotQAQCLevel}
\alias{plotQAQCGap}
\alias{plotQAQCTV}
\title{Plot QAQC Data}
\usage{
plotQAQCLevel(
  x,
  level = c("ol", "tol"),
  dbRange = NULL,
  freqMin = NULL,
  channel = NULL,
  title = NULL
)

plotQAQCGap(x, title = NULL)

plotQAQCTV(x, title = NULL)
}
\arguments{
\item{x}{dataframe created by \link{evaluateDeployment} or
\link{evaluateRecordings}}

\item{level}{which frequency band levels to display, one of "ol"
(octave level) or "tol" (third-octave level).}

\item{dbRange}{range of dB (y-axis) values to plot}

\item{freqMin}{minimum frequency (Hz) to show on plot}

\item{channel}{if \code{NULL}, all channels will be plotted as a
faceted plot, if not then data will be filtered to this channel}

\item{title}{optional title to add to plot}
}
\value{
a ggplot object
}
\description{
Simple plotting functions for various the various types of
quality assurance / quality control (QAQC) data created by
\link{evaluateDeployment} and \link{evaluateRecordings}
}
